requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
six==1.17.0
SQLAlchemy==2.0.20
typing_extensions==4.13.2
Werkzeug==2.3.7
WTForms==3.0.1
